refactor(Search): replacing Search to Input component element
