--- search-films/src/components/Search/index.module.css
@@ -1,25 +1,0 @@
-.search {
-  display: flex;
-}
-.search.vertical {
-  flex-direction: row;
-  gap: 8px;
-}
-.search.horizontal {
-  flex-direction: column;
-  align-items: baseline;
-  gap: 39px;
-}
-
-.icon-search::before {
-    position: absolute;
-    display: block;
-    font-size: inherit;
-    width: 19px;
-    height: 19px;
-    overflow: visible;
-    left: 16px;
-    top: 20px;
-    /* Icon */
-    content: url("/search-icon.svg");
-}
